Add manhole.
We're having issues with processes slowly growing in memory over time
and never releasing it.  I can't seem to reproduce it locally.  It would
be nice to have a way to get in the running process and inspect the
garbage collector to see what objects are being kept around, and what is
referencing them.