Increase minimum size of message before compressing
The total number of websocket processes on the fanout exchanges combined
with r/place's high concurrency is hurting rabbit right now.

This compression threshold was lowered while testing compression and
doesn't need to be so low. We're wasting CPU here for limited gain and
we're currently bottlenecked on CPU not network throughput so by
tweaking this down we can scale down the websocket cluster more to
reduce pressure on rabbit.

The new value is completely arbitrary and we can probably continue up if
we need more headroom, while watching graphs of course.