Add Sentry integration, version 2
The previous attempt used the raven-provided middleware. Unfortunately,
we have to run code in the wsgi handler context as well and it gets
application state from the application so things got rather complex when
the application was wrapped in a middleware. This is a less invasive but
more manual way of adding raven into the mix.